refactor title.  Tests failing, need merge from master
