+++ components/shared/Title/Title.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import { below } from '../../utilities';
+import { below } from '../../../utilities';
 
 const twoColBp = 'large';
